Make Prog and progSeq parameters of BaseItTests
