+++ lms-backend/src/it/scala/scalan/it/lms/MethodCallItTests.scala
@@ -6,93 +6,92 @@ import scala.language.reflectiveCalls
 import scalan.compilation.lms._
 import scalan.compilation.lms.scalac.CommunityLmsCompilerScala
 import scalan.compilation.lms.source2bin.SbtConfig
+import scalan.it.BaseItTests
 import scalan.it.smoke.CommunitySmokeItTests
 import scalan.util.{Exceptions, FileUtil}
 import scalan._
 
-class MethodCallItTests extends CommunitySmokeItTests with BeforeAndAfterAll {
+trait MethodCallTestProg extends ScalanCommunityDsl {
+
+  lazy val emptyIf = fun { (in: Rep[(Boolean, (Double, Double))]) =>
+    val Pair(x, Pair(y, z)) = in
+    IF(x) THEN y ELSE z
+  }
+
+  lazy val exceptionWithIfFalse = fun { (p: Rep[(Int, (SThrowable, SThrowable))]) =>
+    val Pair(n, Pair(t1, t2)) = p
+    IF(n>0) THEN t2.getMessage ELSE t1.initCause(t2).getMessage
+  }
+
+  lazy val exceptionWithIfTrue = fun { (p: Rep[(Int, (SThrowable, SThrowable))]) =>
+    //val Pair(n, Pair(t1, t2)) = p
+    val Pair(n, Pair(t1, t2)) = p
+    IF(n<=0) THEN t2.getMessage ELSE t1.initCause(t2).getMessage
+  }
+
+  lazy val arrayLengthFun = fun { (v: Rep[Array[Int]]) =>
+    //v.arrayLength
+    v.length
+  }
+
+  lazy val arrayOneArg = fun { (in: Rep[(Array[Double], Int)]) =>
+    in._1.apply(in._2)
+  }
+
+  lazy val easyMap = fun { (in: Rep[(Array[Double], (Double, Double))]) =>
+    val Pair(m0, Pair(vFrom, vTo)) = in
+    val m:Rep[Array[Double]] = m0.map(x => x*x)
+    m.reduce
+  }
+
+  lazy val mapWithLambda = fun { (in: Rep[(Array[Double], (Double, Double))]) =>
+    val Pair(m0, Pair(vFrom, vTo)) = in
+    def f(x:Rep[Double],y:Rep[Double], z:Rep[Double]): Rep[Double] = {
+      x*y+z
+    }
+
+    val m:Rep[Array[Double]] = m0.map(x => f(x, vFrom, vTo))
+
+    m.reduce
+  }
+
+  lazy val mapWithLambdaIf = fun { (in: Rep[(Array[Double], (Double, Double))]) =>
+    val Pair(m0, Pair(vFrom, vTo)) = in
+    def f(x:Rep[Double],y:Rep[Double], z:Rep[Double]): Rep[Double] = {
+      IF(x===y) THEN z ELSE x
+    }
+
+    val m:Rep[Array[Double]] = m0.map(x => f(x, vFrom, vTo))
+
+    m.reduce
+  }
+
+  lazy val mapWithLambdaIfGt = fun { (in: Rep[(Array[Double], (Double, Double))]) =>
+    val Pair(m0, Pair(vFrom, vTo)) = in
+    def f(x:Rep[Double],y:Rep[Double], z:Rep[Double]): Rep[Double] = {
+      IF(x<y) THEN x ELSE z
+    }
+
+    val m:Rep[Array[Double]] = m0.map(x => f(x, vFrom, vTo))
+
+    m.reduce
+  }
+
+  lazy val arrayLength = fun { v: Rep[Array[Int]] =>
+    Collection(v).length
+  }
+
+  lazy val message = fun { (t: Rep[String]) => SThrowable(t).getMessage}
+}
+
+class MethodCallItTests extends BaseItTests[MethodCallTestProg](new ScalanCommunityDslSeq with MethodCallTestProg) with BeforeAndAfterAll {
 
   override def beforeAll() = {
     FileUtil.deleteIfExist(prefix)
   }
 
-  trait Prog extends super.Prog {
+  class ProgStaged extends ScalanCommunityDslExp with MethodCallTestProg
 
-    lazy val emptyIf = fun { (in: Rep[(Boolean, (Double, Double))]) =>
-      val Pair(x, Pair(y, z)) = in
-      IF(x) THEN y ELSE z
-    }
-
-    lazy val exceptionWithIfFalse = fun { (p: Rep[(Int, (SThrowable, SThrowable))]) =>
-      val Pair(n, Pair(t1, t2)) = p
-      IF(n>0) THEN t2.getMessage ELSE t1.initCause(t2).getMessage
-    }
-
-    lazy val exceptionWithIfTrue = fun { (p: Rep[(Int, (SThrowable, SThrowable))]) =>
-      //val Pair(n, Pair(t1, t2)) = p
-      val Pair(n, Pair(t1, t2)) = p
-      IF(n<=0) THEN t2.getMessage ELSE t1.initCause(t2).getMessage
-    }
-
-    lazy val arrayLengthFun = fun { (v: Rep[Array[Int]]) =>
-      //v.arrayLength
-      v.length
-    }
-
-    lazy val arrayOneArg = fun { (in: Rep[(Array[Double], Int)]) =>
-      in._1.apply(in._2)
-    }
-
-    lazy val easyMap = fun { (in: Rep[(Array[Double], (Double, Double))]) =>
-      val Pair(m0, Pair(vFrom, vTo)) = in
-      val m:Rep[Array[Double]] = m0.map(x => x*x)
-      m.reduce
-    }
-
-    lazy val mapWithLambda = fun { (in: Rep[(Array[Double], (Double, Double))]) =>
-      val Pair(m0, Pair(vFrom, vTo)) = in
-      def f(x:Rep[Double],y:Rep[Double], z:Rep[Double]): Rep[Double] = {
-        x*y+z
-      }
-
-      val m:Rep[Array[Double]] = m0.map(x => f(x, vFrom, vTo))
-
-      m.reduce
-    }
-
-    lazy val mapWithLambdaIf = fun { (in: Rep[(Array[Double], (Double, Double))]) =>
-      val Pair(m0, Pair(vFrom, vTo)) = in
-      def f(x:Rep[Double],y:Rep[Double], z:Rep[Double]): Rep[Double] = {
-        IF(x===y) THEN z ELSE x
-      }
-
-      val m:Rep[Array[Double]] = m0.map(x => f(x, vFrom, vTo))
-
-      m.reduce
-    }
-
-    lazy val mapWithLambdaIfGt = fun { (in: Rep[(Array[Double], (Double, Double))]) =>
-      val Pair(m0, Pair(vFrom, vTo)) = in
-      def f(x:Rep[Double],y:Rep[Double], z:Rep[Double]): Rep[Double] = {
-        IF(x<y) THEN x ELSE z
-      }
-
-      val m:Rep[Array[Double]] = m0.map(x => f(x, vFrom, vTo))
-
-      m.reduce
-    }
-
-    lazy val arrayLength = fun { v: Rep[Array[Int]] =>
-      Collection(v).length
-    }
-
-    lazy val message = fun { (t: Rep[String]) => SThrowable(t).getMessage}
-  }
-
-  class ProgSeq extends ProgCommunitySeq with Prog {}
-  class ProgStaged extends ProgCommunityExp with Prog {}
-
-  val progSeq = new ProgSeq
   val progStaged = new CommunityLmsCompilerScala(new ProgStaged)
   val defaultCompilers = compilers(progStaged)
